#62 Upravení schémat. Když nastavuji novou hodnotu ve schématu, tak se zkontroluje, zda-li se nová hodnota nerovná se starou hodnotou. Pokud se rovnají, tak není co nastavovat, v opačném případě se nastaví nová hodnota a zavolá se callback, na který lze reagovat v případě změny hodnoty.
