+++ app/src/main/java/cz/zcu/fav/tymsnu/stimulatorremotecontrol/model/AItem.java
@@ -23,4 +23,7 @@ public abstract class AItem {
         this.name = name;
     }
 
+    public interface OnValueChanged {
+        void changed();
+    }
 }
