+++ app/src/main/java/cz/zcu/fav/tymsnu/stimulatorremotecontrol/model/Scheme.java
@@ -91,14 +91,21 @@ public final class Scheme extends AItem {
         return outputCount;
     }
 
-    public void setOutputCount(int outputCount) throws IllegalArgumentException {
+    public void setOutputCount(int outputCount) {setOutputCount(outputCount, null);}
+    public void setOutputCount(int outputCount, OnValueChanged onValueChanged) throws IllegalArgumentException {
         if (outputCount < MIN_OUTPUT_COUNT || outputCount > MAX_OUTPUT_COUNT)
             throw new IllegalArgumentException();
+
+        if (this.outputCount == outputCount)
+            return;
 
         this.outputCount = outputCount;
 
         if (outputList.size() != outputCount)
             rearangeOutputs();
+
+        if (onValueChanged != null)
+            onValueChanged.changed();
     }
 
     public List<Output> getOutputList() {
@@ -109,16 +116,30 @@ public final class Scheme extends AItem {
         return edge;
     }
 
-    public void setEdge(Edge edge) {
+    public void setEdge(Edge edge) {setEdge(edge, null);}
+    public void setEdge(Edge edge, OnValueChanged onValueChanged) {
+        if (this.edge.equals(edge))
+            return;
+
         this.edge = edge;
+
+        if (onValueChanged != null)
+            onValueChanged.changed();
     }
 
     public Random getRandom() {
         return random;
     }
 
-    public void setRandom(Random random) {
+    public void setRandom(Random random) {setRandom(random, null);}
+    public void setRandom(Random random, OnValueChanged onValueChanged) {
+        if (this.random.equals(random))
+            return;
+
         this.random = random;
+
+        if (onValueChanged != null)
+            onValueChanged.changed();
     }
 
     // endregion
@@ -182,4 +203,215 @@ public final class Scheme extends AItem {
         }
     }
     // endregion
+
+    public static final class Output {
+
+        // region Variables
+        // Název výstupu
+        private final String name;
+        // Reference pro nastavení pulsu
+        public final Puls puls;
+        // Reference pro nastavení rozdělení
+        public final Distribution distribution;
+        // Intenzita jasu [%](0-100)
+        private int brightness;
+        // endregion
+
+        // region Constructors
+
+        /**
+         * Konstruktor výstupu
+         * Vytvoří nový výstup s výchozími hodnotami
+         * @param name Název výstupu
+         */
+        public Output(String name) {
+            this(name, new Puls(), new Distribution(), 0);
+        }
+
+        /**
+         * Konstruktor výstupu
+         * Vytvoří nový výstup na zákadě parametrů
+         * @param name Název výstupu
+         * @param puls Reference na nastavení pulsu
+         * @param distribution Reference na nastavení rozdělení
+         * @param brightness Intenzita jasu [%](0-100)
+         */
+        public Output(String name, Puls puls, Distribution distribution, int brightness) {
+            this.name = name;
+            this.puls = puls;
+            this.distribution = distribution;
+            this.brightness = brightness;
+        }
+        // endregion
+
+        // region Private methods
+
+        // endregion
+
+        // region Public methods
+        public boolean canUpdateDistribution(List<Output> outputs, int val) {
+            int sum = 0;
+            for (Output output : outputs) {
+                sum += output.distribution.getValue();
+            }
+            sum -= distribution.getValue();
+
+            return sum + val <= 100;
+
+        }
+        // endregion
+
+        // region Getters & Setters
+
+        /**
+         * Vrátí jméno
+         * @return jméno
+         */
+        public String getName() {
+            return name;
+        }
+
+        /**
+         * Vrátí intenzitu jasu
+         * @return intenzita jasu
+         */
+        public int getBrightness() {
+            return brightness;
+        }
+
+        public void setBrightness(int brightness) {setBrightness(brightness, null);}
+        /**
+         * Nastaví intenzitu jasu
+         * Hodnota musí být v rozmězí <0-100>
+         * Pokud bude hodnota jiná, nic se nenastaví
+         * @param brightness Intenzita jasu
+         */
+        public void setBrightness(int brightness, OnValueChanged onValueChanged) {
+            if (brightness < 0 || brightness > 100)
+                return;
+
+            this.brightness = brightness;
+
+            if (onValueChanged != null)
+                onValueChanged.changed();
+        }
+
+        // endregion
+
+        // region Classes
+        public static final class Puls {
+            // Doba, po kterou jsou výstupy aktivní
+            private int up;
+            // Doba, po kterou jsou výstupy neaktivní
+            private int down;
+
+            /**
+             * Konstruktor pulsu
+             * Vytvoří nový puls s výchozími hodnotami
+             * Up - 0
+             * Down - 0
+             */
+            public Puls() {
+                this(0, 0);
+            }
+
+            /**
+             * Konstruktor pulsu
+             * Vytvoří nový puls na základě parametrů
+             * @param up Doba, po kterou jsou výstupy aktivní
+             * @param down Doba, po kterou jsou výstupy neaktivní
+             */
+            public Puls(int up, int down) {
+                this.up = up;
+                this.down = down;
+            }
+
+            public int getUp() {
+                return up;
+            }
+
+            public void setUp(int up) {setUp(up, null);}
+            public void setUp(int up, OnValueChanged onValueChanged) {
+                if (this.up == up)
+                    return;
+
+                this.up = up;
+
+                if (onValueChanged != null)
+                    onValueChanged.changed();
+            }
+
+            public int getDown() {
+                return down;
+            }
+
+            public void setDown(int down) {setDown(down, null);}
+            public void setDown(int down, OnValueChanged onValueChanged) {
+                if (this.down == down)
+                    return;
+
+                this.down = down;
+
+                if (onValueChanged != null)
+                    onValueChanged.changed();
+            }
+        }
+
+        public static final class Distribution {
+            private int value;
+            private int delay;
+
+            /**
+             * Konstruktor rozdělení
+             * Vytvoří nové rozdělení s výchozími hodnotami
+             * Value - 0
+             * Delay - 0
+             */
+            public Distribution() {
+                this(0, 0);
+            }
+
+            /**
+             * Konstruktor rozdělení
+             * Vytvoří nové rozdělení na základě parametrů
+             * @param value
+             * @param delay
+             */
+            public Distribution(int value, int delay) {
+                this.value = value;
+                this.delay = delay;
+            }
+
+            public int getValue() {
+                return value;
+            }
+
+            public void setValue(int value) {setValue(value, null);}
+            public void setValue(int value, OnValueChanged onValueChanged) {
+                if (this.value == value)
+                    return;
+
+                this.value = value;
+
+                if (onValueChanged != null)
+                    onValueChanged.changed();
+            }
+
+            public int getDelay() {
+                return delay;
+            }
+
+            public void setDelay(int delay) {setDelay(delay, null);}
+            public void setDelay(int delay, OnValueChanged onValueChanged) {
+                if (this.delay == delay)
+                    return;
+
+                this.delay = delay;
+
+                if (onValueChanged != null)
+                    onValueChanged.changed();
+            }
+        }
+        // endregion
+    }
 }
