+++ app/src/main/java/cz/zcu/fav/tymsnu/stimulatorremotecontrol/model/handler/packet/SchemePacketHandler.java
@@ -7,8 +7,8 @@ import cz.zcu.fav.tymsnu.stimulatorremotecontrol.bytes.Codes;
 import cz.zcu.fav.tymsnu.stimulatorremotecontrol.bytes.DataConvertor;
 import cz.zcu.fav.tymsnu.stimulatorremotecontrol.bytes.Packet;
 import cz.zcu.fav.tymsnu.stimulatorremotecontrol.bytes.IPacketable;
-import cz.zcu.fav.tymsnu.stimulatorremotecontrol.model.Output;
 import cz.zcu.fav.tymsnu.stimulatorremotecontrol.model.Scheme;
+import cz.zcu.fav.tymsnu.stimulatorremotecontrol.model.Scheme.Output;
 
 
 public class SchemePacketHandler implements IPacketable {
